Lime2: Increase LDO4 voltage to 3.3 Volts
On the Olimex OLinuXino Lime2 LDO3 and LDO4 are not connected to any
peripheral. Their intended/recommended usage would be for CSI0 and CSI1
and they are thus also connected to the VCC of pin bank E and G
respectively.

This has as a consequence however, that when using pins from bank E and
bank G, we get LDO3 and LDO4 as their power source/reference power,
meaning that with LDO4 at 2.8 for example, UART3 also only operates at
2.8 voltage levels. Worse so when configuring the the LDO4 at 0.0
Volts, the UART will now no longer be able to supply power.

Setting these LDO4 to 3.3 Volt allows ports G to function as expected.

As there are no other peripherals connected to these ports (such as a
CSI based camera) which prefers it's I/O at 2.8 this has no consequence
to the rest of the board.

EM-1631

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>